beam-dev-env/requirements.txt
apache-beam[gcp,aws,azure,test,docs,interactive]==2.53.0
jupyterlab==4.1.2
faker
geocoder
kafka-python